To avoid failure of toggle operation, toggle lower dut to active first, and then shutdown bgpd (#6485)
